Move locale-related classes to Locale namespace
