+++ site/app/Application/Locale/LocaleLink.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types = 1);
 
-namespace MichalSpacekCz\Application;
+namespace MichalSpacekCz\Application\Locale;
 
 class LocaleLink
 {
