+++ site/app/Application/Locale/LocaleLinkGenerator.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types = 1);
 
-namespace MichalSpacekCz\Application;
+namespace MichalSpacekCz\Application\Locale;
 
 use Contributte\Translation\Translator;
 use MichalSpacekCz\Application\Routing\RouterFactory;
