+++ site/app/Www/Presenters/BasePresenter.php
@@ -5,8 +5,8 @@ namespace MichalSpacekCz\Www\Presenters;
 
 use MichalSpacekCz\Application\CriticalCss;
 use MichalSpacekCz\Application\CriticalCssFactory;
-use MichalSpacekCz\Application\LocaleLink;
-use MichalSpacekCz\Application\LocaleLinkGenerator;
+use MichalSpacekCz\Application\Locale\LocaleLink;
+use MichalSpacekCz\Application\Locale\LocaleLinkGenerator;
 use MichalSpacekCz\Application\Theme;
 use MichalSpacekCz\User\Manager;
 use Nette\Application\UI\InvalidLinkException;
