+++ site/app/Www/Presenters/ErrorPresenter.php
@@ -6,8 +6,8 @@ namespace MichalSpacekCz\Www\Presenters;
 use Contributte\Translation\Translator;
 use MichalSpacekCz\Application\AppRequest;
 use MichalSpacekCz\Application\Exceptions\NoOriginalRequestException;
-use MichalSpacekCz\Application\LocaleLink;
-use MichalSpacekCz\Application\LocaleLinkGenerator;
+use MichalSpacekCz\Application\Locale\LocaleLink;
+use MichalSpacekCz\Application\Locale\LocaleLinkGenerator;
 use MichalSpacekCz\EasterEgg\FourOhFourButFound;
 use Nette\Application\BadRequestException;
 use Nette\Application\UI\InvalidLinkException;
