+++ site/config/services.neon
@@ -2,8 +2,8 @@ services:
 	- MichalSpacekCz\Application\AppRequest
 	- MichalSpacekCz\Application\CriticalCssFactory
 	- MichalSpacekCz\Application\Error
-	localeLinkGenerator: MichalSpacekCz\Application\LocaleLinkGenerator(languages: %locales.languages%)
-	- MichalSpacekCz\Application\Locales
+	localeLinkGenerator: MichalSpacekCz\Application\Locale\LocaleLinkGenerator(languages: %locales.languages%)
+	- MichalSpacekCz\Application\Locale\Locales
 	- MichalSpacekCz\Application\Routing\RouterFactory(supportedLocales: %locales.supported%, rootDomainMapping: %locales.rootDomainMapping%, translatedRoutes: %translatedRoutes.presenters%)
 	- @MichalSpacekCz\Application\Routing\RouterFactory::createRouter
 	- MichalSpacekCz\Application\Theme
